requirements.txt
pyautogui>=0.9.54
requests>=2.31.0
Pillow>=10.0.0
pynput>=1.7.6
pytesseract>=0.3.10
